feat: add AnkiConnect plugin support
This commit adds support for uploading cards via AnkiConnect plugin.
Generated cards are now by default uploaded to the Anki via the plugin.
The accompanying images references in the cards are also uploaded via
the plugin.

The addition of the AnkiConnect support means that frontmatter block is
mandatory, since some of the fields are required by the plugin to
upload the cards to the correct Anki deck, to use the correct note type,
etc.

The mandatory formatter fields are:
- deck_name
- note_type_basic
- note_type_cloze

The tags field is optional.

The generation of Anki cards in CSV format is now marked as a legacy
feature, it is by default turned off.

User can also change the default URL of the AnkiConnect server via the
.ini config file.

Future work:
- The code currently skips when trying to upload a card duplicate. In
  the future there an update system could be implemented. This would
  support a usecase where users would like to fix the already uploaded
  card via the markdown input file.
- Currently the users need to provide the images file to the md2anki
  config file. This is not really needed, as the information about the
  image location is in the input markdown document itself. So that could
  be removed and thus UX would be a bit better, since that would be one
  less config option to carry about.
- Currently the images must be local, there is no support for providing
  URL image links. Support for that could be added.

Related: #16
Closes: #3