infra: update versions of GitHub Actions
The changes in this commit were prompted by the deprecation
warning for the actions/upload-artifact@v3, which is used by the
build.yaml for the Python code. One of the PRs failed to pass the CI
because of this warning.

Once I started looking through the workflow files I realized that almost
all actions required the version update.